backend-core: make sendToScorer a replaceable interface
