+++ packages/backend-core/src/domain/articles/countDenormalization.ts
@@ -18,10 +18,10 @@ import {
   Category,
   Comment,
 } from '../../models';
-import { IArticleInstance } from '../../models/article';
+import { IArticleInstance } from '../../models';
 import {
   denormalizeCommentCountsForCategory,
-} from '../categories/countDenormalization';
+} from '../categories';
 
 export async function denormalizeCommentCountsForArticle(article: IArticleInstance | null): Promise<void> {
   if (!article) {
