+++ packages/cli/commands/comments/rescore.ts
@@ -43,7 +43,7 @@ export async function handler() {
 
     for (const c of comments) {
       const comment = (await Comment.findById(c.id))!;
-      await resendForScoring(comment, true);
+      await resendForScoring(comment);
 
       logger.info(`Rescored comment ${c.id}`);
     }
